Fix installation of PR repos, more logging
Previously, installing PR repos with ubports-qa was hit-or-miss. These
changes should finally fix that for good.

I've added a proper Python logger to the script so we aren't printing
things all over the place. This lets us have a '-v' flag for more
verbose logging. I've sprinkled debug logging so it's easier to
troubleshoot the script in the future.